general llm test based on config
