+++ docs/docs/gpt-researcher/llms/testing-your-llm.md
@@ -3,34 +3,27 @@
 Here is a snippet of code to help you verify that your LLM-related environment variables are set up correctly.
 
 ```python
+from gpt_researcher.config.config import Config
+from gpt_researcher.utils.llm import create_chat_completion
 import asyncio
-from gpt_researcher.utils.llm import get_llm
-from gpt_researcher import GPTResearcher
 from dotenv import load_dotenv
 load_dotenv()
 
 async def main():
+    cfg = Config()
 
-    # Example usage of get_llm function
-    llm_provider = "openai"
-    model = "gpt-3.5-turbo" 
-    temperature = 0.7
-    max_tokens = 1000
-
-    llm = get_llm(llm_provider, model=model, temperature=temperature, max_tokens=max_tokens)
-    print(f"LLM Provider: {llm_provider}, Model: {model}, Temperature: {temperature}, Max Tokens: {max_tokens}")
-    print('llm: ',llm)
-    await test_llm(llm=llm)
-
-
-async def test_llm(llm):
-    # Test the connection with a simple query
-    messages = [{"role": "user", "content": "sup?"}]
     try:
-        response = await llm.get_chat_response(messages, stream=False)
-        print("LLM response:", response)
+        report = await create_chat_completion(
+            model=cfg.smart_llm_model,
+            messages = [{"role": "user", "content": "sup?"}],
+            temperature=0.35,
+            llm_provider=cfg.smart_llm_provider,
+            stream=True,
+            max_tokens=cfg.smart_token_limit,
+            llm_kwargs=cfg.llm_kwargs
+        )
     except Exception as e:
-        print(f"Error: {e}")
+        print(f"Error in calling LLM: {e}")
 
 # Run the async function
 asyncio.run(main())
